requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.3
joblib==1.1.0